pyproject.toml
[project]
name = "nldf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "openai-agents>=0.2.3",
    "paho-mqtt>=2.1.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
]
